[Refactor] 프로젝트 작성 페이지 헤더, 제목, 분야, 소개 공통 컴포넌트로 분리 및 관련 이미지 폴더 이동
